Current: changed sensor parameter type from unsigned to uint8_t for memory reasons (accept values only between 0 and 2)

+++ mainboard/Inc/current.h
@@ -51,4 +51,4 @@ current_t *current_get_zero();
 /**
  * @brief Returns the current flowing through the specified sensor
  */
-current_t current_get_current_from_sensor(unsigned sensor);
+current_t current_get_current_from_sensor(uint8_t sensor);
